# openapi.yaml
openapi: 3.0.0
info:
  description: Define an API
  title: Restful API
  version: 1.0.0
paths:
...
  "/{sql}/update/{tablename}":
    put:
      tags:
        - Table CRUD
      summary: 更新資料表數值
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: peoples
        - description: 更新的欄位數值(colname,value)
          name: set
          in: query
          schema:
            type: array
            example: ["key1,123456"]
            items:
              type: string
          style: form
          explode: true
        - description: 被更新的欄位條件(colname,value)
          name: where
          in: query
          schema:
            type: array
            example: ["key2,223456"]
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                type: string
              example: 
                  "Successfully Update Value"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
  "/jointable/{sql1}/{table1}/{sql2}/{table2}":
    get:
      tags:
        - Join Table
      summary: 資料表合併
      parameters:
        - description: 資料庫引擎
          name: sql1
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: table1
          in: path
          required: true
          schema:
            type: string
          example: peoples
        - description: 資料庫引擎
          name: sql2
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: table2
          in: path
          required: true
          schema:
            type: string
          example: events
        - description: join欄位
          name: join
          in: query
          required: true
          schema:
            type: array
            example: ["key1", "key2"]
            items:
              type: string
        - description: table1挑選欄位
          name: table1
          in: query
          schema:
            type: array
            example: ["bith"]
            items:
              type: string
        - description: table2挑選欄位
          name: table2
          in: query
          schema:
            type: array
            example: ["position"]
            items:
              type: string
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example: 
                uuid: "專屬uuid"
                欄位名稱: "欄位數值"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
servers:
  - url: http://localhost:8080/v1
components:
  schemas:
...
    models.Error:
      type: object
      properties:
        message:
          type: string
    models.object:
      type: object